requirements.txt
py3Fdfs==2.2.0
JPype1==1.2.1